Group similar functions into their own namespaces. SHA224 and SHA256 are identical in every aspect except their initial H values and their digest size.
